Add a custom View for emoji questions
This now manages setting up the views (as well as any future work).
This results in the users of this class only having to deal with forwarding
the actual question to the view rather than setting it up.
There is now *one* place where all set-up happens rather than multiple ones
with the same code.