Add chat answers to all questions
Messages can already be answered. However, the actual answer was not stored
in any way. Answering would only cause a message to be sent.
Each instance of Question now has an answer property
(of type Any?, is there a better way to do this?). This is now set when a message
is answered.