Add time zones to activities file
This should make changes in the user's time zone a bit less bad.
This obviously won't fix changes in the user's time zone during an activity.
Another solution is needed for that.
This change wasn't made as an update to the UserActivityStorage version
(it's still 1) because the app is still in development and no users will
be affected by this.