Add max value to app usage chart
This is an experimental change that seems to work fine.
The left axis already has a spacing of 25f (= 25% based on the documentation).
However, this wasn't enough as bars could still go through the roof.
This should now be fixed, as the chart's max value is now 1.25x the largest value.