Add pomodoro exercise
While this is not final, it's mostly completed.
This adds an exercise guiding the user through the Pomodoro technique.

Navigation and storage of this are two things that are not fully handled
right now. However, they will be once more plans arrive.

The strings are still only written in German, this needs to be changed in the future.

This commit may be a bit too large, sorry :).