Add swipe options to the goal lists in the Eisenhower Matrix
The user can now perform a swipe on each goal in each list to show options to
delete and pin said goals. This is also saved to the file.
This implementation now follows what's shown in the app's UI concept, though it's
still not completely finished. See below notes for issues that are still present.

This was one of the most frustrating experiences I've ever had implementing something.
It literally took 6 or 7 days (!) to add this because there were various issues
too complicated to describe here. I've added some documentation so it can
hopefully be understood more easily in the future.

Issues:
- The cards' shadows aren't consistent. They sometimes overlap each other.
- A card's shadow will disappear if there's only one card in the list.
  This probably has something to do with the issue above.
- Open menus aren't closed when other ones are opened. This can cause numerous
  issues and is probably related to the shadow issues as well.
  Menus are also not closed when scrolling.
- This implementation doesn't always show animations. Closing menus can sometimes
  just happen without an animation.
- Scrolling an item out of the visible area with its menu open will cause the
  menu to disappear.