Add completion status to goals
Goals can now be completed by checking their CheckBox or by adding them
as completed. This will be saved to the file as well.

I didn't update the version of the goal storage format because there are no
public devices running this version of the app and an update would be pointless
when there's no data being lost.