Prompt user for info on startup whenever necessary
The name input screen and the profile screen will be shown whenever it is
needed by the app. For example, if the user didn't enter any info
(like on the app's first start) both the name input screen and the profile screen
to enter more info will be shown. However, when - as an example - only the email
address is needed, the app will show the profile screen. The user can then enter
their info and only then continue to the rest of the app.