Fix bottom sheet header not rendering correctly
This issue only occurred when the layout was expanded and the user returned
to the start screen from another tab. Because fragments aren't reused properly yet,
the value of maxDistanceFromTop was reset. This caused a division by 0, leading to
a NaN value for expansionValue. This then caused the method to return before
rendering anything.
maxDistanceFromTop is now calculated in a better way that shouldn't break
anymore.