Fix some interactive chat messages not working
While trying to take screenshots, I found out that you couldn't
actually answer multiple choice, picture, emoji, and slider
questions in the app's chat. This is now fixed.

The problem was this: I added a default behavior for these views
(they're dedicated views, that's why text answers still worked)
some time ago. These came in the form of listeners added to
components of these views. These listeners were set once a
question was bound to the views. The chat fragment added its own
listener a bit earlier, however.
This meant that the right listener was set and was replaced with
the wrong (the default) one a bit later, leading to this bug.